+++ app/sim_bench_test.go
@@ -7,12 +7,14 @@ import (
 	"os"
 	"testing"
 
+	"github.com/stretchr/testify/require"
+
 	cmtproto "github.com/cometbft/cometbft/proto/tendermint/types"
 
+	"github.com/cosmos/cosmos-sdk/baseapp"
 	simtestutil "github.com/cosmos/cosmos-sdk/testutil/sims"
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
 	"github.com/cosmos/cosmos-sdk/x/simulation"
-	simcli "github.com/cosmos/cosmos-sdk/x/simulation/client/cli"
 )
 
 // Profile with:
@@ -20,31 +22,31 @@ import (
 func BenchmarkFullAppSimulation(b *testing.B) {
 	b.ReportAllocs()
 	config, db, dir, logger, skip, err := setupSimulation("goleveldb-app-sim", "Simulation")
-	if err != nil {
-		b.Fatalf("simulation setup failed: %s", err.Error())
-	}
+	require.NoError(b, err, "simulation setup failed")
 
 	if skip {
 		b.Skip("skipping benchmark application simulation")
 	}
-	PrintConfig(config)
+	printConfig(config)
 
 	defer func() {
-		db.Close()
-		err = os.RemoveAll(dir)
-		if err != nil {
-			b.Fatal(err)
-		}
+		require.NoError(b, db.Close())
+		require.NoError(b, os.RemoveAll(dir))
 	}()
 
-	app := New(logger, db, nil, true, map[int64]bool{}, b.TempDir(), simcli.FlagPeriodValue, simtestutil.EmptyAppOptions{}, interBlockCacheOpt())
+	appOpts := newSimAppOpts(b)
+	baseAppOpts := []func(*baseapp.BaseApp){
+		interBlockCacheOpt(),
+		baseapp.SetChainID(config.ChainID),
+	}
+	app := New(logger, db, nil, true, appOpts, baseAppOpts...)
 
 	// run randomized simulation
 	_, simParams, simErr := simulation.SimulateFromSeed(
 		b,
 		os.Stdout,
 		app.BaseApp,
-		ProvAppStateFn(app.AppCodec(), app.SimulationManager(), app.DefaultGenesis()),
+		provAppStateFn(app.AppCodec(), app.SimulationManager(), app.DefaultGenesis()),
 		simtypes.RandomAccounts, // Replace with own random account function if using keys other than secp256k1
 		simtestutil.SimulationOperations(app, app.AppCodec(), config),
 		app.ModuleAccountAddrs(),
@@ -53,47 +55,43 @@ func BenchmarkFullAppSimulation(b *testing.B) {
 	)
 
 	// export state and simParams before the simulation error is checked
-	if err = simtestutil.CheckExportSimulation(app, config, simParams); err != nil {
-		b.Fatal(err)
-	}
+	err = simtestutil.CheckExportSimulation(app, config, simParams)
+	require.NoError(b, err, "CheckExportSimulation")
+	require.NoError(b, simErr, "SimulateFromSeed")
 
-	if simErr != nil {
-		b.Fatal(simErr)
-	}
-
-	PrintStats(config, db)
+	printStats(config, db)
 }
 
 func BenchmarkInvariants(b *testing.B) {
 	b.ReportAllocs()
 	config, db, dir, logger, skip, err := setupSimulation("leveldb-app-invariant-bench", "Simulation")
-	if err != nil {
-		b.Fatalf("simulation setup failed: %s", err.Error())
-	}
+	require.NoError(b, err, "simulation setup failed")
 
 	if skip {
 		b.Skip("skipping benchmark application simulation")
 	}
 
 	config.AllInvariants = false
-	PrintConfig(config)
+	printConfig(config)
 
 	defer func() {
-		db.Close()
-		err = os.RemoveAll(dir)
-		if err != nil {
-			b.Fatal(err)
-		}
+		require.NoError(b, db.Close())
+		require.NoError(b, os.RemoveAll(dir))
 	}()
 
-	app := New(logger, db, nil, true, map[int64]bool{}, b.TempDir(), simcli.FlagPeriodValue, simtestutil.EmptyAppOptions{}, interBlockCacheOpt())
+	appOpts := newSimAppOpts(b)
+	baseAppOpts := []func(*baseapp.BaseApp){
+		interBlockCacheOpt(),
+		baseapp.SetChainID(config.ChainID),
+	}
+	app := New(logger, db, nil, true, appOpts, baseAppOpts...)
 
 	// run randomized simulation
 	_, lastBlockTime, simParams, simErr := simulation.SimulateFromSeedProv(
 		b,
 		os.Stdout,
 		app.BaseApp,
-		ProvAppStateFn(app.AppCodec(), app.SimulationManager(), app.DefaultGenesis()),
+		provAppStateFn(app.AppCodec(), app.SimulationManager(), app.DefaultGenesis()),
 		simtypes.RandomAccounts, // Replace with own random account function if using keys other than secp256k1
 		simtestutil.SimulationOperations(app, app.AppCodec(), config),
 		app.ModuleAccountAddrs(),
@@ -102,15 +100,11 @@ func BenchmarkInvariants(b *testing.B) {
 	)
 
 	// export state and simParams before the simulation error is checked
-	if err = simtestutil.CheckExportSimulation(app, config, simParams); err != nil {
-		b.Fatal(err)
-	}
+	err = simtestutil.CheckExportSimulation(app, config, simParams)
+	require.NoError(b, err, "CheckExportSimulation")
+	require.NoError(b, simErr, "SimulateFromSeedProv")
 
-	if simErr != nil {
-		b.Fatal(simErr)
-	}
-
-	PrintStats(config, db)
+	printStats(config, db)
 
 	ctx := app.NewContextLegacy(true, cmtproto.Header{Height: app.LastBlockHeight() + 1, Time: lastBlockTime})