+++ app/sim_test.go
@@ -3,6 +3,7 @@ package app
 // DONTCOVER
 
 import (
+	"bytes"
 	"encoding/json"
 	"fmt"
 	"math/rand"
@@ -13,6 +14,8 @@ import (
 	"testing"
 	"time"
 
+	wasmtypes "github.com/CosmWasm/wasmd/x/wasm/types"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	abci "github.com/cometbft/cometbft/abci/types"
@@ -21,42 +24,29 @@ import (
 	"cosmossdk.io/log"
 	"cosmossdk.io/store"
 	storetypes "cosmossdk.io/store/types"
-	evidencetypes "cosmossdk.io/x/evidence/types"
-
-	icqtypes "github.com/cosmos/ibc-apps/modules/async-icq/v8/types"
+	"cosmossdk.io/x/feegrant"
 
 	dbm "github.com/cosmos/cosmos-db"
 	"github.com/cosmos/cosmos-sdk/baseapp"
+	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/codec"
+	"github.com/cosmos/cosmos-sdk/server"
 	simtestutil "github.com/cosmos/cosmos-sdk/testutil/sims"
+	"github.com/cosmos/cosmos-sdk/types/kv"
 	"github.com/cosmos/cosmos-sdk/types/module"
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
-	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	authzkeeper "github.com/cosmos/cosmos-sdk/x/authz/keeper"
-	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
-	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
-	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
-	minttypes "github.com/cosmos/cosmos-sdk/x/mint/types"
-	paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
 	"github.com/cosmos/cosmos-sdk/x/simulation"
 	simcli "github.com/cosmos/cosmos-sdk/x/simulation/client/cli"
 	slashingtypes "github.com/cosmos/cosmos-sdk/x/slashing/types"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
-	capabilitytypes "github.com/cosmos/ibc-go/modules/capability/types"
+	"github.com/cosmos/gogoproto/proto"
+	icqtypes "github.com/cosmos/ibc-apps/modules/async-icq/v8/types"
 	icagenesistypes "github.com/cosmos/ibc-go/v8/modules/apps/27-interchain-accounts/genesis/types"
 	icatypes "github.com/cosmos/ibc-go/v8/modules/apps/27-interchain-accounts/types"
 
 	cmdconfig "github.com/provenance-io/provenance/cmd/provenanced/config"
 	"github.com/provenance-io/provenance/internal/pioconfig"
-	attributetypes "github.com/provenance-io/provenance/x/attribute/types"
-	"github.com/provenance-io/provenance/x/hold"
-	markertypes "github.com/provenance-io/provenance/x/marker/types"
-	metadatatypes "github.com/provenance-io/provenance/x/metadata/types"
-	msgfeetype "github.com/provenance-io/provenance/x/msgfees/types"
-	nametypes "github.com/provenance-io/provenance/x/name/types"
-	"github.com/provenance-io/provenance/x/quarantine"
-	"github.com/provenance-io/provenance/x/sanction"
-	triggertypes "github.com/provenance-io/provenance/x/trigger/types"
 )
 
 func init() {
@@ -64,64 +54,161 @@ func init() {
 	pioconfig.SetProvenanceConfig("", 0)
 }
 
-type StoreKeysPrefixes struct {
-	A        storetypes.StoreKey
-	B        storetypes.StoreKey
-	Prefixes [][]byte
-}
-
-// ProvAppStateFn wraps the simtypes.AppStateFn and sets the ICA GenesisState if isn't yet defined in the appState.
-func ProvAppStateFn(cdc codec.JSONCodec, simManager *module.SimulationManager, genesisState map[string]json.RawMessage) simtypes.AppStateFn {
+// provAppStateFn wraps the simtypes.AppStateFn and sets the ICA and ICQ GenesisState if isn't yet defined in the appState.
+func provAppStateFn(cdc codec.JSONCodec, simManager *module.SimulationManager, genesisState map[string]json.RawMessage) simtypes.AppStateFn {
 	return func(r *rand.Rand, accs []simtypes.Account, config simtypes.Config) (json.RawMessage, []simtypes.Account, string, time.Time) {
 		appState, simAccs, chainID, genesisTimestamp := simtestutil.AppStateFn(cdc, simManager, genesisState)(r, accs, config)
 		appState = appStateWithICA(appState, cdc)
 		appState = appStateWithICQ(appState, cdc)
+		appState = appStateWithWasmSeqs(appState, cdc)
 		return appState, simAccs, chainID, genesisTimestamp
 	}
 }
 
 // appStateWithICA checks the given appState for an ica entry. If it's not found, it's populated with the defaults.
 func appStateWithICA(appState json.RawMessage, cdc codec.JSONCodec) json.RawMessage {
-	rawState := make(map[string]json.RawMessage)
-	err := json.Unmarshal(appState, &rawState)
-	if err != nil {
-		panic(fmt.Sprintf("error unmarshalling appstate: %v", err))
-	}
-	icaGenJSON, icaGenFound := rawState[icatypes.ModuleName]
-	if !icaGenFound || len(icaGenJSON) == 0 {
-		icaGenState := icagenesistypes.DefaultGenesis()
-		rawState[icatypes.ModuleName] = cdc.MustMarshalJSON(icaGenState)
-		appState, err = json.Marshal(rawState)
-		if err != nil {
-			panic(fmt.Sprintf("error marshalling appstate: %v", err))
-		}
-	}
-	return appState
+	return mutateAppState(appState, cdc, icatypes.ModuleName, icagenesistypes.DefaultGenesis(), nil)
 }
 
 // appStateWithICA checks the given appState for an ica entry. If it's not found, it's populated with the defaults.
 func appStateWithICQ(appState json.RawMessage, cdc codec.JSONCodec) json.RawMessage {
+	return mutateAppState(appState, cdc, icqtypes.ModuleName, icqtypes.DefaultGenesis(), nil)
+}
+
+// appStateWithWasmSeqs ensures the wasm genesis state has sequence entries.
+func appStateWithWasmSeqs(appState json.RawMessage, cdc codec.JSONCodec) json.RawMessage {
+	// During the import/export test, the wasm module state of the first app ends up with empty sequence entries.
+	// But export-genesis reads empty sequence entries as "1" and creates an entry in the Sequences list.
+	// So when we call InitGenesis with the second app, those sequences are stored in state, resulting in different states between the two apps.
+	// The behavior of the wasm module is the same if the value is 1 or empty, so the behavior of the two apps is the same, just not the state.
+	// So in here, we make sure there are default entries for the sequences so that even the first one has them.
+	// If the sims do any wasm stuff that'll change those entries, both the state and exported genesis will match, and there won't be a problem.
+	return mutateAppState(appState, cdc, wasmtypes.ModuleName, &wasmtypes.GenesisState{Params: wasmtypes.DefaultParams()}, func(wasmGenState *wasmtypes.GenesisState) *wasmtypes.GenesisState {
+		requiredSequences := [][]byte{wasmtypes.KeySequenceCodeID, wasmtypes.KeySequenceInstanceID}
+		for _, seqKey := range requiredSequences {
+			if !sequenceExists(wasmGenState.Sequences, seqKey) {
+				wasmGenState.Sequences = append(wasmGenState.Sequences, wasmtypes.Sequence{IDKey: seqKey, Value: 1})
+			}
+		}
+		return wasmGenState
+	})
+}
+
+// sequenceExists checks if a sequence entry exists for the given key.
+func sequenceExists(sequences []wasmtypes.Sequence, key []byte) bool {
+	for _, seq := range sequences {
+		if bytes.Equal(seq.IDKey, key) {
+			return true
+		}
+	}
+	return false
+}
+
+// mutateAppState returns a new appState with an updated (or new) entry for the given module.
+// The mutator will receive the module's genesis state from appState if it exists, otherwise,
+// it'll receive the provided defaultState. If the mutator is nil, this basically just ensures
+// that the appState has an entry for the module, setting it to the default if it's not already there.
+func mutateAppState[G proto.Message](appState json.RawMessage, cdc codec.JSONCodec, moduleName string, defaultState G, mutator func(state G) G) json.RawMessage {
 	rawState := make(map[string]json.RawMessage)
 	err := json.Unmarshal(appState, &rawState)
 	if err != nil {
 		panic(fmt.Sprintf("error unmarshalling appstate: %v", err))
 	}
-	icqGenJSON, icqGenFound := rawState[icqtypes.ModuleName]
-	if !icqGenFound || len(icqGenJSON) == 0 {
-		icqGenState := icqtypes.DefaultGenesis()
-		rawState[icqtypes.ModuleName] = cdc.MustMarshalJSON(icqGenState)
-		appState, err = json.Marshal(rawState)
+
+	if len(rawState[moduleName]) > 0 {
+		err = cdc.UnmarshalJSON(rawState[moduleName], defaultState)
 		if err != nil {
-			panic(fmt.Sprintf("error marshalling appstate: %v", err))
+			panic(fmt.Sprintf("error unmarshalling %s genesis state from %q as %T: %v",
+				moduleName, string(rawState[moduleName]), defaultState, err))
 		}
 	}
+
+	if mutator != nil {
+		defaultState = mutator(defaultState)
+	}
+
+	rawState[moduleName], err = cdc.MarshalJSON(defaultState)
+	if err != nil {
+		panic(fmt.Sprintf("error marshalling %s genesis state %#v: %v",
+			moduleName, defaultState, err))
+	}
+
+	appState, err = json.Marshal(rawState)
+	if err != nil {
+		panic(fmt.Sprintf("error marshalling appstate: %v", err))
+	}
+
 	return appState
 }
 
 func setupSimulation(dirPrefix string, dbName string) (simtypes.Config, dbm.DB, string, log.Logger, bool, error) {
 	config := simcli.NewConfigFromFlags()
+	config.ChainID = pioconfig.SimAppChainID
 	db, dir, logger, skip, err := simtestutil.SetupSimulation(config, dirPrefix, dbName, simcli.FlagVerboseValue, simcli.FlagEnabledValue)
 	return config, db, dir, logger, skip, err
+}
+
+// fauxMerkleModeOpt returns a BaseApp option to use a dbStoreAdapter instead of
+// an IAVLStore for faster simulation speed.
+func fauxMerkleModeOpt(bapp *baseapp.BaseApp) {
+	bapp.SetFauxMerkleMode()
+}
+
+// interBlockCacheOpt returns a BaseApp option function that sets the persistent
+// inter-block write-through cache.
+func interBlockCacheOpt() func(*baseapp.BaseApp) {
+	return baseapp.SetInterBlockCache(store.NewCommitKVStoreCacheManager())
+}
+
+// printStats outputs the config and db info.
+func printStats(config simtypes.Config, db dbm.DB) {
+	printConfig(config)
+	if config.Commit {
+		printDBInfo(db)
+	}
+}
+
+// printConfig outputs the config.
+func printConfig(config simtypes.Config) {
+	fmt.Println("-vvv-  Config Info  -vvv-")
+	cfields := cmdconfig.MakeFieldValueMap(config, true)
+	for _, f := range cfields.GetSortedKeys() {
+		fmt.Printf("%s: %s\n", f, cfields.GetStringOf(f))
+	}
+	fmt.Println("-^^^-  Config Info  -^^^-")
+}
+
+// printDBInfo outputs the db.Stats map.
+func printDBInfo(db dbm.DB) {
+	fmt.Println("-vvv-  Database Info  -vvv-")
+	dbStats := db.Stats()
+	if len(dbStats) == 0 {
+		fmt.Println("No info to report.")
+	} else {
+		keys := make([]string, 0, len(dbStats))
+		for k := range dbStats {
+			keys = append(keys, k)
+		}
+		sort.Strings(keys)
+		for _, k := range keys {
+			v := dbStats[k]
+			if strings.Contains(v, "\n") {
+				fmt.Printf("%s:\n", k)
+				fmt.Println(v)
+			} else {
+				fmt.Printf("%s: %q\n", k, v)
+			}
+		}
+	}
+	fmt.Println("-^^^-  Database Info  -^^^-")
+}
+
+// newSimAppOpts creates a new set of AppOptions with a temp dir for home, and the desired invariant check period.
+func newSimAppOpts(t testing.TB) simtestutil.AppOptionsMap {
+	return simtestutil.AppOptionsMap{
+		flags.FlagHome:            t.TempDir(),
+		server.FlagInvCheckPeriod: simcli.FlagPeriodValue,
+	}
 }
 
 func TestFullAppSimulation(t *testing.T) {
@@ -129,15 +216,20 @@ func TestFullAppSimulation(t *testing.T) {
 	if skip {
 		t.Skip("skipping provenance application simulation")
 	}
-	PrintConfig(config)
+	printConfig(config)
 	require.NoError(t, err, "provenance simulation setup failed")
 
 	defer func() {
-		db.Close()
+		require.NoError(t, db.Close())
 		require.NoError(t, os.RemoveAll(dir))
 	}()
 
-	app := New(logger, db, nil, true, map[int64]bool{}, t.TempDir(), simcli.FlagPeriodValue, simtestutil.EmptyAppOptions{}, fauxMerkleModeOpt)
+	appOpts := newSimAppOpts(t)
+	baseAppOpts := []func(*baseapp.BaseApp){
+		fauxMerkleModeOpt,
+		baseapp.SetChainID(config.ChainID),
+	}
+	app := New(logger, db, nil, true, appOpts, baseAppOpts...)
 	require.Equal(t, "provenanced", app.Name())
 
 	fmt.Printf("running provenance full app simulation\n")
@@ -147,7 +239,7 @@ func TestFullAppSimulation(t *testing.T) {
 		t,
 		os.Stdout,
 		app.BaseApp,
-		ProvAppStateFn(app.AppCodec(), app.SimulationManager(), app.DefaultGenesis()),
+		provAppStateFn(app.AppCodec(), app.SimulationManager(), app.DefaultGenesis()),
 		simtypes.RandomAccounts, // Replace with own random account function if using keys other than secp256k1
 		simtestutil.SimulationOperations(app, app.AppCodec(), config),
 		app.ModuleAccountAddrs(),
@@ -157,10 +249,10 @@ func TestFullAppSimulation(t *testing.T) {
 
 	// export state and simParams before the simulation error is checked
 	err = simtestutil.CheckExportSimulation(app, config, simParams)
-	require.NoError(t, err)
-	require.NoError(t, simErr)
+	require.NoError(t, err, "CheckExportSimulation")
+	require.NoError(t, simErr, "SimulateFromSeed")
 
-	PrintStats(config, db)
+	printStats(config, db)
 }
 
 func TestSimple(t *testing.T) {
@@ -168,15 +260,20 @@ func TestSimple(t *testing.T) {
 	if skip {
 		t.Skip("skipping provenance application simulation")
 	}
-	PrintConfig(config)
+	printConfig(config)
 	require.NoError(t, err, "provenance simulation setup failed")
 
 	defer func() {
-		db.Close()
+		require.NoError(t, db.Close())
 		require.NoError(t, os.RemoveAll(dir))
 	}()
 
-	app := New(logger, db, nil, true, map[int64]bool{}, t.TempDir(), simcli.FlagPeriodValue, simtestutil.EmptyAppOptions{}, fauxMerkleModeOpt)
+	appOpts := newSimAppOpts(t)
+	baseAppOpts := []func(*baseapp.BaseApp){
+		fauxMerkleModeOpt,
+		baseapp.SetChainID(config.ChainID),
+	}
+	app := New(logger, db, nil, true, appOpts, baseAppOpts...)
 	require.Equal(t, "provenanced", app.Name())
 
 	// run randomized simulation
@@ -184,7 +281,7 @@ func TestSimple(t *testing.T) {
 		t,
 		os.Stdout,
 		app.BaseApp,
-		ProvAppStateFn(app.AppCodec(), app.SimulationManager(), app.DefaultGenesis()),
+		provAppStateFn(app.AppCodec(), app.SimulationManager(), app.DefaultGenesis()),
 		simtypes.RandomAccounts, // Replace with own random account function if using keys other than secp256k1
 		simtestutil.SimulationOperations(app, app.AppCodec(), config),
 		app.ModuleAccountAddrs(),
@@ -192,8 +289,8 @@ func TestSimple(t *testing.T) {
 		app.AppCodec(),
 	)
 
-	require.NoError(t, simErr)
-	PrintStats(config, db)
+	require.NoError(t, simErr, "SimulateFromSeed")
+	printStats(config, db)
 }
 
 // Profile with:
@@ -201,12 +298,22 @@ func TestSimple(t *testing.T) {
 func TestAppImportExport(t *testing.T) {
 	// uncomment to run in ide without flags.
 	//simcli.FlagEnabledValue = true
+	//tempDir, err := os.MkdirTemp("", "sim-log-*")
+	//require.NoError(t, err, "MkdirTemp")
+	//t.Logf("tempDir: %s", tempDir)
+	//simcli.FlagNumBlocksValue = 30
+	//simcli.FlagVerboseValue = true
+	//simcli.FlagCommitValue = true
+	//simcli.FlagSeedValue = 2
+	//simcli.FlagPeriodValue = 3
+	//simcli.FlagExportParamsPathValue = filepath.Join(tempDir, fmt.Sprintf("sim_params-%d.json", simcli.FlagSeedValue))
+	//simcli.FlagExportStatePathValue = filepath.Join(tempDir, fmt.Sprintf("sim_state-%d.json", simcli.FlagSeedValue))
 
 	config, db, dir, logger, skip, err := setupSimulation("leveldb-app-sim", "Simulation")
 	if skip {
 		t.Skip("skipping application import/export simulation")
 	}
-	PrintConfig(config)
+	printConfig(config)
 	require.NoError(t, err, "simulation setup failed")
 
 	defer func() {
@@ -214,8 +321,13 @@ func TestAppImportExport(t *testing.T) {
 		require.NoError(t, os.RemoveAll(dir))
 	}()
 
-	home := t.TempDir()
-	app := New(logger, db, nil, true, map[int64]bool{}, home, simcli.FlagPeriodValue, simtestutil.EmptyAppOptions{}, fauxMerkleModeOpt)
+	appOpts := newSimAppOpts(t)
+	baseAppOpts := []func(*baseapp.BaseApp){
+		fauxMerkleModeOpt,
+		baseapp.SetChainID(config.ChainID),
+	}
+	app := New(logger, db, nil, true, appOpts, baseAppOpts...)
+	require.Equal(t, "provenanced", app.Name())
 
 	fmt.Printf("running provenance test import export\n")
 
@@ -224,7 +336,7 @@ func TestAppImportExport(t *testing.T) {
 		t,
 		os.Stdout,
 		app.BaseApp,
-		ProvAppStateFn(app.AppCodec(), app.SimulationManager(), app.DefaultGenesis()),
+		provAppStateFn(app.AppCodec(), app.SimulationManager(), app.DefaultGenesis()),
 		simtypes.RandomAccounts, // Replace with own random account function if using keys other than secp256k1
 		simtestutil.SimulationOperations(app, app.AppCodec(), config),
 		app.ModuleAccountAddrs(),
@@ -234,87 +346,93 @@ func TestAppImportExport(t *testing.T) {
 
 	// export state and simParams before the simulation error is checked
 	err = simtestutil.CheckExportSimulation(app, config, simParams)
-	require.NoError(t, err)
-	require.NoError(t, simErr)
+	require.NoError(t, err, "CheckExportSimulation")
+	require.NoError(t, simErr, "SimulateFromSeedProv")
 
-	PrintStats(config, db)
+	printStats(config, db)
 
 	fmt.Printf("exporting genesis...\n")
 
 	exported, err := app.ExportAppStateAndValidators(false, nil, nil)
-	require.NoError(t, err)
+	require.NoError(t, err, "ExportAppStateAndValidators")
 
 	fmt.Printf("importing genesis...\n")
 
-	_, newDB, newDir, _, _, err := setupSimulation("leveldb-app-sim-2", "Simulation-2")
-	require.NoError(t, err, "simulation setup failed")
+	newDB, newDir, newLogger, _, err := simtestutil.SetupSimulation(config, "leveldb-app-sim-2", "Simulation-2", simcli.FlagVerboseValue, simcli.FlagEnabledValue)
+	require.NoError(t, err, "simulation setup 2 failed")
 
 	defer func() {
 		require.NoError(t, newDB.Close())
 		require.NoError(t, os.RemoveAll(newDir))
 	}()
 
-	newApp := New(log.NewNopLogger(), newDB, nil, true, map[int64]bool{}, home, simcli.FlagPeriodValue, simtestutil.EmptyAppOptions{}, fauxMerkleModeOpt)
+	newApp := New(newLogger, newDB, nil, true, appOpts, baseAppOpts...)
 
 	var genesisState map[string]json.RawMessage
 	err = json.Unmarshal(exported.AppState, &genesisState)
 	require.NoError(t, err)
 
-	defer func() {
-		if r := recover(); r != nil {
-			rstr := fmt.Sprintf("%v", r)
-			if !strings.Contains(rstr, "validator set is empty after InitGenesis") {
-				panic(r)
-			}
-			logger.Info("Skipping simulation as all validators have been unbonded")
-			logger.Info("err", rstr, "stacktrace", string(debug.Stack()))
-		}
-	}()
-
 	ctxA := app.NewContextLegacy(true, cmtproto.Header{Height: app.LastBlockHeight(), Time: lastBlockTime})
 	ctxB := newApp.NewContextLegacy(true, cmtproto.Header{Height: app.LastBlockHeight(), Time: lastBlockTime})
-	newApp.mm.InitGenesis(ctxB, app.AppCodec(), genesisState)
-	newApp.StoreConsensusParams(ctxB, exported.ConsensusParams)
+	_, err = newApp.mm.InitGenesis(ctxB, app.AppCodec(), genesisState)
+	if err != nil {
+		if strings.Contains(err.Error(), "validator set is empty after InitGenesis") {
+			logger.Info("Skipping simulation as all validators have been unbonded")
+			logger.Info("err", err, "stacktrace", string(debug.Stack()))
+			return
+		}
+	}
+	require.NoError(t, err, "InitGenesis")
+
+	err = newApp.StoreConsensusParams(ctxB, exported.ConsensusParams)
+	require.NoError(t, err, "StoreConsensusParams")
 
 	fmt.Printf("comparing stores...\n")
 
-	storeKeysPrefixes := []StoreKeysPrefixes{
-		{app.keys[authtypes.StoreKey], newApp.keys[authtypes.StoreKey], [][]byte{}},
-		{app.keys[stakingtypes.StoreKey], newApp.keys[stakingtypes.StoreKey],
-			[][]byte{
-				stakingtypes.UnbondingQueueKey, stakingtypes.RedelegationQueueKey, stakingtypes.ValidatorQueueKey,
-				stakingtypes.HistoricalInfoKey,
-			}}, // ordering may change but it doesn't matter
-		{app.keys[slashingtypes.StoreKey], newApp.keys[slashingtypes.StoreKey], [][]byte{}},
-		{app.keys[minttypes.StoreKey], newApp.keys[minttypes.StoreKey], [][]byte{}},
-		{app.keys[distrtypes.StoreKey], newApp.keys[distrtypes.StoreKey], [][]byte{}},
-		{app.keys[banktypes.StoreKey], newApp.keys[banktypes.StoreKey], [][]byte{banktypes.BalancesPrefix}},
-		{app.keys[paramtypes.StoreKey], newApp.keys[paramtypes.StoreKey], [][]byte{}},
-		{app.keys[govtypes.StoreKey], newApp.keys[govtypes.StoreKey], [][]byte{}},
-		{app.keys[evidencetypes.StoreKey], newApp.keys[evidencetypes.StoreKey], [][]byte{}},
-		{app.keys[capabilitytypes.StoreKey], newApp.keys[capabilitytypes.StoreKey], [][]byte{}},
-		{app.keys[authzkeeper.StoreKey], newApp.keys[authzkeeper.StoreKey], [][]byte{authzkeeper.GrantKey, authzkeeper.GrantQueuePrefix}},
-		{app.keys[quarantine.StoreKey], newApp.keys[quarantine.StoreKey], [][]byte{}},
-		{app.keys[sanction.StoreKey], newApp.keys[sanction.StoreKey], [][]byte{}},
-
-		{app.keys[markertypes.StoreKey], newApp.keys[markertypes.StoreKey], [][]byte{}},
-		{app.keys[msgfeetype.StoreKey], newApp.keys[msgfeetype.StoreKey], [][]byte{}},
-		{app.keys[attributetypes.StoreKey], newApp.keys[attributetypes.StoreKey], [][]byte{attributetypes.AttributeAddrLookupKeyPrefix}},
-		{app.keys[nametypes.StoreKey], newApp.keys[nametypes.StoreKey], [][]byte{}},
-		{app.keys[metadatatypes.StoreKey], newApp.keys[metadatatypes.StoreKey], [][]byte{}},
-		{app.keys[triggertypes.StoreKey], newApp.keys[triggertypes.StoreKey], [][]byte{}},
-		{app.keys[hold.StoreKey], newApp.keys[hold.StoreKey], [][]byte{}},
+	// skip certain prefixes
+	skipPrefixes := map[string][][]byte{
+		stakingtypes.StoreKey: {
+			stakingtypes.UnbondingQueueKey, stakingtypes.RedelegationQueueKey, stakingtypes.ValidatorQueueKey,
+			stakingtypes.HistoricalInfoKey, stakingtypes.UnbondingIDKey, stakingtypes.UnbondingIndexKey,
+			stakingtypes.UnbondingTypeKey, stakingtypes.ValidatorUpdatesKey,
+		},
+		authzkeeper.StoreKey:   {authzkeeper.GrantQueuePrefix},
+		feegrant.StoreKey:      {feegrant.FeeAllowanceQueueKeyPrefix},
+		slashingtypes.StoreKey: {slashingtypes.ValidatorMissedBlockBitmapKeyPrefix},
+		wasmtypes.StoreKey:     {wasmtypes.TXCounterPrefix},
 	}
 
-	for _, skp := range storeKeysPrefixes {
-		storeA := ctxA.KVStore(skp.A)
-		storeB := ctxB.KVStore(skp.B)
+	storeKeys := app.GetStoreKeys()
+	require.NotEmpty(t, storeKeys, "storeKeys")
 
-		failedKVAs, failedKVBs := simtestutil.DiffKVStores(storeA, storeB, skp.Prefixes)
-		require.Equal(t, len(failedKVAs), len(failedKVBs), "unequal sets of key-values to compare")
+	for _, appKeyA := range storeKeys {
+		keyName := appKeyA.Name()
+		t.Run(keyName, func(t *testing.T) {
+			// only compare kvstores
+			if _, ok := appKeyA.(*storetypes.KVStoreKey); !ok {
+				t.Skipf("Skipping because the key is a %T (not a KVStoreKey)", appKeyA)
+				return
+			}
 
-		fmt.Printf("compared %d different key/value pairs between %s and %s\n", len(failedKVAs), skp.A, skp.B)
-		require.Equal(t, 0, len(failedKVAs), simtestutil.GetSimulationLog(skp.A.Name(), app.SimulationManager().StoreDecoders, failedKVAs, failedKVBs))
+			appKeyB := newApp.GetKey(keyName)
+
+			storeA := ctxA.KVStore(appKeyA)
+			storeB := ctxB.KVStore(appKeyB)
+
+			failedKVAs, failedKVBs := simtestutil.DiffKVStores(storeA, storeB, skipPrefixes[keyName])
+			assert.Equal(t, len(failedKVAs), len(failedKVBs), "unequal sets of key-values to compare: %s", keyName)
+			fmt.Printf("compared %d different key/value pairs between %s and %s\n", len(failedKVAs), appKeyA, appKeyB)
+
+			// Make the lists the same length because GetSimulationLog assumes they're that way.
+			for len(failedKVBs) < len(failedKVAs) {
+				failedKVBs = append(failedKVBs, kv.Pair{Key: []byte{}, Value: []byte{}})
+			}
+			for len(failedKVBs) > len(failedKVAs) {
+				failedKVAs = append(failedKVAs, kv.Pair{Key: []byte{}, Value: []byte{}})
+			}
+
+			assert.Equal(t, 0, len(failedKVAs), simtestutil.GetSimulationLog(keyName, app.SimulationManager().StoreDecoders, failedKVAs, failedKVBs))
+		})
 	}
 }
 
@@ -323,23 +441,27 @@ func TestAppSimulationAfterImport(t *testing.T) {
 	if skip {
 		t.Skip("skipping application simulation after import")
 	}
-	PrintConfig(config)
+	printConfig(config)
 	require.NoError(t, err, "simulation setup failed")
 
 	defer func() {
-		db.Close()
+		require.NoError(t, db.Close())
 		require.NoError(t, os.RemoveAll(dir))
 	}()
 
-	home := t.TempDir()
-	app := New(logger, db, nil, true, map[int64]bool{}, home, simcli.FlagPeriodValue, simtestutil.EmptyAppOptions{}, fauxMerkleModeOpt)
+	appOpts := newSimAppOpts(t)
+	baseAppOpts := []func(*baseapp.BaseApp){
+		fauxMerkleModeOpt,
+		baseapp.SetChainID(config.ChainID),
+	}
+	app := New(logger, db, nil, true, appOpts, baseAppOpts...)
 
 	// Run randomized simulation
 	stopEarly, lastBlockTime, simParams, simErr := simulation.SimulateFromSeedProv(
 		t,
 		os.Stdout,
 		app.BaseApp,
-		ProvAppStateFn(app.AppCodec(), app.SimulationManager(), app.DefaultGenesis()),
+		provAppStateFn(app.AppCodec(), app.SimulationManager(), app.DefaultGenesis()),
 		simtypes.RandomAccounts, // Replace with own random account function if using keys other than secp256k1
 		simtestutil.SimulationOperations(app, app.AppCodec(), config),
 		app.ModuleAccountAddrs(),
@@ -349,10 +471,10 @@ func TestAppSimulationAfterImport(t *testing.T) {
 
 	// export state and simParams before the simulation error is checked
 	err = simtestutil.CheckExportSimulation(app, config, simParams)
-	require.NoError(t, err)
-	require.NoError(t, simErr)
+	require.NoError(t, err, "CheckExportSimulation")
+	require.NoError(t, simErr, "SimulateFromSeedProv")
 
-	PrintStats(config, db)
+	printStats(config, db)
 
 	if stopEarly {
 		fmt.Println("can't export or import a zero-validator genesis, exiting test...")
@@ -362,22 +484,23 @@ func TestAppSimulationAfterImport(t *testing.T) {
 	fmt.Printf("exporting genesis...\n")
 
 	exported, err := app.ExportAppStateAndValidators(true, nil, nil)
-	require.NoError(t, err)
+	require.NoError(t, err, "ExportAppStateAndValidators")
 
 	fmt.Printf("importing genesis...\n")
 
-	_, newDB, newDir, _, _, err := setupSimulation("leveldb-app-sim-2", "Simulation-2")
-	require.NoError(t, err, "simulation setup failed")
+	newDB, newDir, newLogger, _, err := simtestutil.SetupSimulation(config, "leveldb-app-sim-2", "Simulation-2", simcli.FlagVerboseValue, simcli.FlagEnabledValue)
+	require.NoError(t, err, "simulation setup 2 failed")
 
 	defer func() {
-		newDB.Close()
+		require.NoError(t, newDB.Close())
 		require.NoError(t, os.RemoveAll(newDir))
 	}()
 
-	newApp := New(log.NewNopLogger(), newDB, nil, true, map[int64]bool{}, home, simcli.FlagPeriodValue, simtestutil.EmptyAppOptions{}, fauxMerkleModeOpt)
+	newApp := New(newLogger, newDB, nil, true, appOpts, baseAppOpts...)
 
 	_, err = newApp.InitChain(&abci.RequestInitChain{
 		AppStateBytes: exported.AppState,
+		ChainId:       config.ChainID,
 		Time:          lastBlockTime,
 	})
 	require.NoError(t, err, "InitChain")
@@ -387,7 +510,7 @@ func TestAppSimulationAfterImport(t *testing.T) {
 		t,
 		os.Stdout,
 		newApp.BaseApp,
-		ProvAppStateFn(app.AppCodec(), app.SimulationManager(), app.DefaultGenesis()),
+		provAppStateFn(app.AppCodec(), app.SimulationManager(), app.DefaultGenesis()),
 		simtypes.RandomAccounts, // Replace with own random account function if using keys other than secp256k1
 		simtestutil.SimulationOperations(newApp, newApp.AppCodec(), config),
 		app.ModuleAccountAddrs(),
@@ -422,19 +545,27 @@ func TestAppStateDeterminism(t *testing.T) {
 	numTimesToRunPerSeed := 5
 	appHashList := make([]json.RawMessage, numTimesToRunPerSeed)
 
-	home := t.TempDir()
-
-	seeds := make([]int64, numSeeds)
-	for i := range seeds {
-		seeds[i] = rand.Int63()
+	var seeds []int64
+	if config.Seed != simcli.DefaultSeedValue {
+		// If a seed was provided, just do that one.
+		numSeeds = 1
+		seeds = append(seeds, config.Seed)
+	} else {
+		// Otherwise, pick random seeds to use.
+		seeds = make([]int64, numSeeds)
+		for i := range seeds {
+			seeds[i] = rand.Int63()
+		}
 	}
 
-	// uncomment and tweak to use a single specific seed.
-	//seeds = []int64{9171851189930047994}
+	appOpts := newSimAppOpts(t)
+	if simcli.FlagVerboseValue {
+		appOpts[flags.FlagLogLevel] = "debug"
+	}
 
 	for i, seed := range seeds {
 		config.Seed = seed
-		PrintConfig(config)
+		printConfig(config)
 
 		for j := 0; j < numTimesToRunPerSeed; j++ {
 			var logger log.Logger
@@ -445,7 +576,7 @@ func TestAppStateDeterminism(t *testing.T) {
 			}
 
 			db := dbm.NewMemDB()
-			app := New(logger, db, nil, true, map[int64]bool{}, home, simcli.FlagPeriodValue, simtestutil.EmptyAppOptions{}, interBlockCacheOpt())
+			app := New(logger, db, nil, true, appOpts, interBlockCacheOpt(), baseapp.SetChainID(config.ChainID))
 
 			fmt.Printf(
 				"running provenance non-determinism simulation; seed %d: %d/%d, attempt: %d/%d\n",
@@ -456,7 +587,7 @@ func TestAppStateDeterminism(t *testing.T) {
 				t,
 				os.Stdout,
 				app.BaseApp,
-				ProvAppStateFn(app.AppCodec(), app.SimulationManager(), app.DefaultGenesis()),
+				provAppStateFn(app.AppCodec(), app.SimulationManager(), app.DefaultGenesis()),
 				simtypes.RandomAccounts, // Replace with own random account function if using keys other than secp256k1
 				simtestutil.SimulationOperations(app, app.AppCodec(), config),
 				app.ModuleAccountAddrs(),
@@ -465,7 +596,7 @@ func TestAppStateDeterminism(t *testing.T) {
 			)
 			require.NoError(t, err)
 
-			PrintStats(config, db)
+			printStats(config, db)
 
 			appHash := app.LastCommitID().Hash
 			appHashList[j] = appHash
@@ -478,59 +609,4 @@ func TestAppStateDeterminism(t *testing.T) {
 			}
 		}
 	}
-}
-
-// fauxMerkleModeOpt returns a BaseApp option to use a dbStoreAdapter instead of
-// an IAVLStore for faster simulation speed.
-func fauxMerkleModeOpt(bapp *baseapp.BaseApp) {
-	bapp.SetFauxMerkleMode()
-}
-
-// interBlockCacheOpt returns a BaseApp option function that sets the persistent
-// inter-block write-through cache.
-func interBlockCacheOpt() func(*baseapp.BaseApp) {
-	return baseapp.SetInterBlockCache(store.NewCommitKVStoreCacheManager())
-}
-
-// PrintStats outputs the config and db info.
-func PrintStats(config simtypes.Config, db dbm.DB) {
-	PrintConfig(config)
-	if config.Commit {
-		PrintDBInfo(db)
-	}
-}
-
-// PrintConfig outputs the config.
-func PrintConfig(config simtypes.Config) {
-	fmt.Println("-vvv-  Config Info  -vvv-")
-	cfields := cmdconfig.MakeFieldValueMap(config, true)
-	for _, f := range cfields.GetSortedKeys() {
-		fmt.Printf("%s: %s\n", f, cfields.GetStringOf(f))
-	}
-	fmt.Println("-^^^-  Config Info  -^^^-")
-}
-
-// PrintDBInfo outputs the db.Stats map.
-func PrintDBInfo(db dbm.DB) {
-	fmt.Println("-vvv-  Database Info  -vvv-")
-	dbStats := db.Stats()
-	if len(dbStats) == 0 {
-		fmt.Println("No info to report.")
-	} else {
-		keys := make([]string, 0, len(dbStats))
-		for k := range dbStats {
-			keys = append(keys, k)
-		}
-		sort.Strings(keys)
-		for _, k := range keys {
-			v := dbStats[k]
-			if strings.Contains(v, "\n") {
-				fmt.Printf("%s:\n", k)
-				fmt.Println(v)
-			} else {
-				fmt.Printf("%s: %q\n", k, v)
-			}
-		}
-	}
-	fmt.Println("-^^^-  Database Info  -^^^-")
 }