+++ CHANGELOG.md
@@ -102,6 +102,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * Implement the ProposalMsgs module interface for the internal/provwasm, ibcratelimit, oracle, and sanction modules [#1993](https://github.com/provenance-io/provenance/pull/1993).
 * Update ibcnet relayer to v2.5.2 and fix chain ids for localnet and ibcnet [#2021](https://github.com/provenance-io/provenance/pull/2021).
 * Breakdown internal/helpers into multiple internal packages [#2019](https://github.com/provenance-io/provenance/pull/2019).
+* Update `app.New` to get the home directory, invariant check period, and skip-upgrade heights from the appOptions instead of arguments [#2015](https://github.com/provenance-io/provenance/pull/2015).
+* Simplify the module lists (e.g. `SetOrderEndBlockers`) by removing unneeded entries [#2015](https://github.com/provenance-io/provenance/pull/2015).
 
 ### Client Breaking
 
@@ -114,6 +116,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### Bug Fixes
 
 * The `add-net-asset-values` command now correctly uses the from `flag`'s `AccAddress` [#1995](https://github.com/provenance-io/provenance/issues/1995).
+* Fix the sim tests [#2015](https://github.com/provenance-io/provenance/pull/2015).
 
 ### Deprecated